# filedossier-api/target/generated-sources/openapi-spec/openapi.json
{
  "openapi" : "3.0.1",
  "paths" : {
    "/dossiers/{dossierKey}/{dossierPackage}/{dossierCode}" : {
      "get" : {
        "operationId" : "getDossier",
        "parameters" : [ {
          "name" : "dossierKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierPackage",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/dossier"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/dossier"
                }
              },
              "application/xhtml+xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/dossier"
                }
              }
            }
          }
        }
      }
    },
    "/dossiers/{dossierKey}/{dossierPackage}/{dossierCode}/dossierfiles/{fileCode}" : {
      "get" : {
        "operationId" : "getContents",
        "parameters" : [ {
          "name" : "fileCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierPackage",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/vnd.oasis.opendocument.spreadsheet" : { },
              "application/pdf" : { }
            }
          }
        }
      },
      "put" : {
        "operationId" : "setContents",
        "parameters" : [ {
          "name" : "fileCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierPackage",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "*/*" : { }
            }
          }
        }
      },
      "post" : {
        "operationId" : "uploadContents",
        "parameters" : [ {
          "name" : "fileCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierPackage",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "*/*" : { }
            }
          }
        }
      }
    },
    "/dossiers/{dossierKey}/{dossierPackage}/{dossierCode}/dossierfiles/{fileCode}/context" : {
      "get" : {
        "operationId" : "getContext",
        "parameters" : [ {
          "name" : "fileCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierPackage",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JsonMapObject"
                }
              }
            }
          }
        }
      },
      "put" : {
        "operationId" : "setContext",
        "parameters" : [ {
          "name" : "fileCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierPackage",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dossierCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/JsonMapObject"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "*/*" : { }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "dossier" : {
        "required" : [ "code", "name", "valid" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "valid" : {
            "type" : "string"
          },
          "dossierFile" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/dossierFile"
            }
          }
        },
        "xml" : {
          "name" : "dossier"
        }
      },
      "dossierFile" : {
        "required" : [ "code", "name" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "exists" : {
            "type" : "boolean"
          },
          "readonly" : {
            "type" : "boolean"
          },
          "required" : {
            "type" : "boolean"
          }
        },
        "xml" : {
          "name" : "dossierFile"
        }
      },
      "JsonMapObject" : {
        "type" : "object",
        "properties" : {
          "updateCount" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        }
      }
    }
  }
}