# filedossier-api/target/generated-sources/openapi/src/model/JsonMapObject.js
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 *
 */

import ApiClient from '../ApiClient';

/**
 * The JsonMapObject model module.
 * @module model/JsonMapObject
 * @version 1.0-SNAPSHOT
 */
class JsonMapObject {
    /**
     * Constructs a new <code>JsonMapObject</code>.
     * @alias module:model/JsonMapObject
     */
    constructor() { 
        
        JsonMapObject.initialize(this);
    }

    /**
     * Initializes the fields of this object.
     * This method is used by the constructors of any subclasses, in order to implement multiple inheritance (mix-ins).
     * Only for internal use.
     */
    static initialize(obj) { 
    }

    /**
     * Constructs a <code>JsonMapObject</code> from a plain JavaScript object, optionally creating a new instance.
     * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
     * @param {Object} data The plain JavaScript object bearing properties of interest.
     * @param {module:model/JsonMapObject} obj Optional instance to populate.
     * @return {module:model/JsonMapObject} The populated <code>JsonMapObject</code> instance.
     */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new JsonMapObject();

            if (data.hasOwnProperty('updateCount')) {
                obj['updateCount'] = ApiClient.convertToType(data['updateCount'], {'String': Object});
            }
        }
        return obj;
    }


}

/**
 * @member {Object.<String, Object>} updateCount
 */
JsonMapObject.prototype['updateCount'] = undefined;






export default JsonMapObject;